+++ Runtime/ScriptableEvents/Events/VoidEvent.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace DapperDino.DapperTools.ScriptableEvents.Events
+{
+    [CreateAssetMenu(fileName = "New Void Event", menuName = "Game Events/Void Event")]
+    public class VoidEvent : BaseGameEvent<Void>
+    {
+        public void Raise() => Raise(new Void());
+    }
+}