+++ Runtime/ScriptableEvents/Listeners/IGameEventListener.cs
@@ -1,0 +1,7 @@
+﻿namespace DapperDino.DapperTools.ScriptableEvents.Listeners
+{
+    public interface IGameEventListener<T>
+    {
+        void OnEventRaised(T item);
+    }
+}