+++ Runtime/ScriptableEvents/Listeners/VoidListener.cs
@@ -1,0 +1,7 @@
+﻿using DapperDino.DapperTools.ScriptableEvents.Events;
+using DapperDino.DapperTools.ScriptableEvents.UnityEvents;
+
+namespace DapperDino.DapperTools.ScriptableEvents.Listeners
+{
+    public class VoidListener : BaseGameEventListener<Void, VoidEvent, UnityVoidEvent> { }
+}