+++ Runtime/ScriptableEvents/UnityEvents/UnityItemEvent.cs
@@ -1,0 +1,7 @@
+﻿using System;
+using UnityEngine.Events;
+
+namespace DapperDino.DapperTools.ScriptableEvents.UnityEvents
+{
+    [Serializable] public class UnityVoidEvent : UnityEvent<Void> { }
+}