+++ Runtime/ScriptableEvents/Void.cs
@@ -1,0 +1,4 @@
+﻿namespace DapperDino.DapperTools.ScriptableEvents
+{
+    public struct Void { }
+}