+++ CHANGELOG.md
@@ -41,3 +41,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Component extensions (I currently see no use as they were just copies of the GameObject extensions)
+
+## [0.4.0] - 2019-11-16
+### Added
+- ScriptableObject events
+- First reusable component (timer)
+- Tests for the timer component